+++ tests/Sabre/DAV/ServerPropsInfiniteDepthTest.php
@@ -35,16 +35,10 @@ class ServerPropsInfiniteDepthTest extends AbstractServer {
 
     private function sendRequest($body) {
 
-        $serverVars = array(
-            'REQUEST_URI'    => '/',
-            'REQUEST_METHOD' => 'PROPFIND',
-            'HTTP_DEPTH'          => 'infinity',
-        );
-
-        $request = HTTP\Sapi::createFromServerArray($serverVars);
+        $request = new HTTP\Request('PROPFIND', '/', ['Depth' => 'infinity']);
         $request->setBody($body);
 
-        $this->server->httpRequest = ($request);
+        $this->server->httpRequest = $request;
         $this->server->exec();
 
     }
@@ -66,9 +60,8 @@ class ServerPropsInfiniteDepthTest extends AbstractServer {
 
         $this->sendRequest("");
 
+        $this->assertEquals(207, $this->response->status, 'Incorrect status received. Full response body: ' . $this->response->getBodyAsString());
         $this->assertTrue($hasFired);
-
-        $this->assertEquals(207, $this->response->status);
 
         $this->assertEquals(array(
                 'Content-Type' => 'application/xml; charset=utf-8',